Reworked rankings so they are completely agnostic to firebase

+++ app/src/main/java/ch/epfl/sdp/peakar/rankings/RankingItem.java
@@ -7,14 +7,14 @@ package ch.epfl.sdp.peakar.rankings;
 public class RankingItem {
     private String uid;
     private String username;
-    private int points;
+    private long points;
 
     /**
      * Constructor
      * @param username of user.
      * @param points user has.
      */
-    public RankingItem(String uid, String username, int points) {
+    public RankingItem(String uid, String username, long points) {
         this.uid = uid;
         this.username = username;
         this.points = points;
@@ -40,7 +40,7 @@ public class RankingItem {
      * Getter of the ranking item user points
      * @return points of the user
      */
-    public int getPoints() {
+    public long getPoints() {
         return points;
     }
 }
